+++ include/core/ZipFile.h
@@ -200,7 +200,7 @@ namespace jm
           \param offset The offset in the data buffer where writing should start.
           \param length The number of bytes to write.
           */
-          void write(uint8* data, Integer offset, Integer length);
+          void write(uint8* data, int64 offset, int64 length);
 
           /*!
           \brief Writes the content of the file to the stream and closes the file.
